Use BlobOutput for captured images and add grab
BlobOutput is now used for images, which makes it easier to
get the image from the action.
I've also added grab image, which pulls from the MJPEG stream.
In the future, perhaps this could be provided "for free" by the
stream descriptor - but that currently is confusing, because
you'd end up with actions that don't correspond to methods.