Suppress small changes in exposure_time
When making repeated full resolution captures, exposure_time
was decreased by 20us each time.

When making lots of captures (like in a scan) this became a significant
problem as the exposure started drifting.

I've now added a tolerance to `update_persistent_controls`
that will stop this: if the value read is within a tolerance (defined as 30us) we won't bother updating the stored
value.

This will need improving if we are to work at very short
exposures, because at that point a change of 20us will
become significant - but for now, it solves the problem.