Fixed the tuning test
Deliberately using a bad tuning file was breaking the tuning test - I couldn't
recreate Picamera objects.

I've now updated the default tuning
such that it fetches the camera's type at
import time - this means I can keep
resetting the tuning, even if the camera
is in an odd state.

It may be worth moving this logic into the
main module as well.